Separate tracer utils and data
