+++ SpatialGDK/Source/SpatialGDK/Public/Interop/Connection/SpatialWorkerConnection.h
@@ -48,11 +48,7 @@ public:
 	PhysicalWorkerName GetWorkerId() const;
 	const TArray<FString>& GetWorkerAttributes() const;
 
-	DECLARE_MULTICAST_DELEGATE_OneParam(FOnEnqueueMessage, const SpatialGDK::FOutgoingMessage*);
-	FOnEnqueueMessage OnEnqueueMessage;
-
-	DECLARE_MULTICAST_DELEGATE_OneParam(FOnDequeueMessage, const SpatialGDK::FOutgoingMessage*);
-	FOnDequeueMessage OnDequeueMessage;
+	void BindLatencyTracer(SpatialGDK::TracerSharedPtr Tracer);
 
 	void QueueLatestOpList();
 	void ProcessOutgoingMessages();
@@ -87,4 +83,7 @@ private:
 
 	// Coordinates the async worker ops thread.
 	TOptional<WorkerConnectionCoordinator> ThreadWaitCondition;
+
+	// Used to coordinate logging tracer events
+	SpatialGDK::TracerSharedPtr Tracer;
 };
