+++ SpatialGDK/Source/SpatialGDK/Public/Utils/ComponentFactory.h
@@ -5,6 +5,7 @@
 #include "Interop/SpatialClassInfoManager.h"
 #include "Schema/Interest.h"
 #include "Utils/RepDataUtils.h"
+#include "Utils/SpatialLatencyTracerData.h"
 
 #include <WorkerSDK/improbable/c_schema.h>
 #include <WorkerSDK/improbable/c_worker.h>
@@ -14,7 +15,6 @@ DECLARE_LOG_CATEGORY_EXTERN(LogComponentFactory, Log, All);
 class USpatialNetDriver;
 class USpatialPackageMap;
 class USpatialClassInfoManager;
-class USpatialLatencyTracer;
 class USpatialPackageMapClient;
 
 class UNetDriver;
@@ -28,7 +28,7 @@ namespace SpatialGDK
 class SPATIALGDK_API ComponentFactory
 {
 public:
-	ComponentFactory(bool bInterestDirty, USpatialNetDriver* InNetDriver, USpatialLatencyTracer* LatencyTracer);
+	ComponentFactory(bool bInterestDirty, USpatialNetDriver* InNetDriver, SpatialGDK::TracerSharedPtr LatencyTracer);
 
 	TArray<FWorkerComponentData> CreateComponentDatas(UObject* Object, const FClassInfo& Info, const FRepChangeState& RepChangeState, const FHandoverChangeState& HandoverChangeState, uint32& OutBytesWritten);
 	TArray<FWorkerComponentUpdate> CreateComponentUpdates(UObject* Object, const FClassInfo& Info, Worker_EntityId EntityId, const FRepChangeState* RepChangeState, const FHandoverChangeState* HandoverChangeState, uint32& OutBytesWritten);
@@ -55,7 +55,7 @@ private:
 
 	bool bInterestHasChanged;
 
-	USpatialLatencyTracer* LatencyTracer;
+	SpatialGDK::TracerSharedPtr LatencyTracer;
 };
 
 } // namespace SpatialGDK
